[perl #115830] Fix crash by not copying DESTROY cache
See commits 8c34e50dc and 7cc6787e9db (the latter was reverted by
95f9781bc).

If we are going to store a DESTROY cache in SvSTASH when !SvOBJECT, we
have to make sure not to copy that cache, otherwise other threads will
try to call a CV in the first thread, resulting in a crash.

8c34e50dc introduced this bug, but the crash didn’t start happening
until 7cc6787e9db, and only on Windows.