Stop cv-to-glob assignment redef warnings from leaking
We should not increment the reference count of the assignee until we
are past the warnings.  It should only happen when the reference-
counted location is actually made to hold the assignee.

At the same time, I changed it to use a more specific variant of
SvREFCNT_inc, for speed.