Consider /... a directory component with EFS on VMS.
For some reason the omnibus patch (360732b5267d5dfe) that brought
Extended Filename Syntax support to the VMS port considered three
dots part of the filename portion rather than part of the
directory portion when converting a Unix-format path to VMS format.

There's no reason this should be different with EFS, so make it
the same as without EFS, which gets two TODO tests passing.