subst.t: Test something I nearly broke
The saving and restoring of $@ when utf8 tables were looked up was
causing pp_subst’s string pointers to go stale.  The existing mechanism
of copying it temporarily to another scalar was already very fragile,
and the new COW mechanism broke it (but I fixed it in the same commit
that introduced the new COW mechanism).