Remove 3 unused interpreter variables
These variables have been unused in the Perl core since
commit 4c88d5e0740d796bf5064336d280bba72897f385.

The variables are undocumented.  The only real use of any of these I
found in CPAN is at
https://metacpan.org/source/ABERGMAN/Devel-GC-Helper-0.25/Helper.xs#L1
The uses there appear to be in a list of known Perl variables.  Since
the module was published, more than a few new variables have been added,
making this code obsolete anyway.