mktables: Sort some outputs for repeatability
The recent change to random hash ordering caused some of the files
output by mktables to vary from run to run.  Everything still worked.
However, one of the ways I debug mktables is to make a change, and then
compare the tables it generates with those from before the change.  That
tells me the precise effect of the change.  That no longer works if the
tables come out in random order from run to run.

This patch just sorts certain things so that the tables are output in
the same order each time.