regcomp.c: Collapse cases in a switch()
[:upper:] and [:lower:] have the same logic as some other cases in the
switch statement, but differ in that under /i they match more than just
themselves, and this has to be accounted for.

By moving the test for /i to outside the switch(), these cases can be
collapsed.  There is a small performance penalty in having to test for
all classes if /i is active, and if so, if the class is one of these
two.